Removing templates from abstract matrices and vectors, also factory matrices
