+++ src/matrices/ac_matrix.cpp
@@ -1,7 +1,7 @@
 // IN THIS FILE: Implementation of an abstract class to represent
 // matrices
 
-#include "ac_matrix.tpl.h"
+#include "ac_matrix.h"
 
 namespace scicellxx
 {
@@ -9,24 +9,21 @@ namespace scicellxx
  // ===================================================================
  // Empty constructor
  // ===================================================================
- template<class T>
- ACMatrix<T>::ACMatrix() 
+ ACMatrix<Real>::ACMatrix() 
   : NRows(0), NColumns(0), Is_own_memory_allocated(false), Delete_matrix(true)
  { }
  
  // ===================================================================
  // Constructor to create an m X n zero matrix
  // ===================================================================
- template<class T>
- ACMatrix<T>::ACMatrix(const unsigned long m, const unsigned long n)
+ ACMatrix<Real>::ACMatrix(const unsigned long m, const unsigned long n)
   : NRows(m), NColumns(n), Is_own_memory_allocated(false), Delete_matrix(true)
  { }
  
  // ===================================================================
  // Destructor
  // ===================================================================
- template<class T>
- ACMatrix<T>::~ACMatrix()
+ ACMatrix<Real>::~ACMatrix()
  { }
  
 }
